+++ pkg/repositories/experiment.go
@@ -7,8 +7,8 @@ import (
 	"github.com/rotisserie/eris"
 	"gorm.io/gorm"
 
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/dao/models"
 	"github.com/G-Research/fasttrackml/pkg/database"
+	"github.com/G-Research/fasttrackml/pkg/models"
 )
 
 // ExperimentRepositoryProvider provides an interface to work with `experiment` entity.
